Register toolbar in custom site form
