+++ components/layout/OpenAuthoringSiteForm.tsx
@@ -8,6 +8,7 @@ import {
 import { useEffect } from 'react'
 
 import { useCMS } from 'tinacms'
+import Cookies from 'js-cookie'
 interface Props extends InlineFormProps {
   editMode: boolean
   previewError?: string
@@ -30,6 +31,47 @@ const OpenAuthoringSiteForm = ({
      */
     setTimeout(() => (cms.sidebar.hidden = !editMode), 1)
   }, [])
+
+  /**
+   * Toolbar Plugins
+   */
+  useEffect(() => {
+    const plugins = [
+      {
+        __type: 'toolbar:tool',
+        name: 'create-pr',
+        component: () => <button>Pull Request</button>,
+      },
+      {
+        __type: 'toolbar:status',
+        name: 'current-fork',
+        component: () => <div>{Cookies.get('fork_full_name')}</div>,
+      },
+      {
+        __type: 'toolbar:form-actions',
+        name: 'base-form-actions',
+        component: () => (
+          <>
+            <button onClick={form.reset}>Reset</button>
+            <button onClick={form.submit}>Save</button>
+          </>
+        ),
+      },
+    ] as any
+
+    const removePlugins = () => {
+      plugins.forEach(plugin => cms.plugins.remove(plugin))
+    }
+
+    if (editMode) {
+      plugins.forEach(plugin => cms.plugins.add(plugin))
+    } else {
+      removePlugins()
+    }
+
+    return removePlugins
+  }, [editMode, form])
+
   return (
     <InlineForm
       form={form}
